Etch-A-Sketch Project
I am not very good at updating my git with this project.  I completed the project and it seemed mostly functional except a few things were broken. Realized I could write a function for some of the reoccuring lines in my JS. Instead of deleting, I created a JS.2 file and rewrote. I needed some help, I am not comfortable with arrays and nodelists yet. But nonetheless, This is fairly functional. Still need to add the 'Color' option as per the project instructions.